Added some data integrity checks, also fixed ambassador functionality - fixed an issue where the wrong decks were fed into some strategy methods

+++ controller/controller.go
@@ -7,7 +7,6 @@ import (
   "github.com/twoodhouse/coup-sim/model/table"
   "github.com/twoodhouse/coup-sim/model/player"
 	"strconv"
-	// "github.com/twoodhouse/coup-sim/model/strategies/noLieStrategy"
 )
 
 func StartGame(strategies []strategy.Interface, playerNames []string, numGames int) {
@@ -98,7 +97,6 @@ func DoTurn(player *player.Entity, otherPlayers []*player.Entity, log *log.Entit
 	//top-level challenge logic - it can only happen with these actions
 	if (action == "tax" || action == "steal" || action == "assassinate" || action == "exchange") && !player.Dead(){
 		for i := 0; i < len(otherPlayers); i++ {
-			fmt.Printf("Challenge Checking - Number of other players: %d\n", len(otherPlayers))
 			if !otherPlayers[i].Dead() && otherPlayers[i].Strategy().GetChallenge(log, table.PlayerNames(), table.PlayerCoins(), table.FaceupDecks(), otherPlayers[i].Deck()) {
 				challengeSuccess := !player.Deck().HasCardForAction(action)
 				losingPlayer := player
@@ -107,7 +105,20 @@ func DoTurn(player *player.Entity, otherPlayers []*player.Entity, log *log.Entit
 					losingPlayer = player
 				} else {
 					losingPlayer = otherPlayers[i]
-					swapChallengedCard(player, table, log, 1)
+					var cardValue int
+					switch action {
+					case "tax":
+						cardValue = 1
+					case "steal":
+						cardValue = 2
+					case "assassinate":
+						cardValue = 3
+					case "block":
+						cardValue = 4
+					case "exchange":
+						cardValue = 5
+					}
+					swapChallengedCard(player, table, log, cardValue)
 				}
 				log.CreateChallenge(otherPlayers[i].Name(), challengeSuccess)
 				cardLoss := revealCard(losingPlayer, table, log)
@@ -118,12 +129,15 @@ func DoTurn(player *player.Entity, otherPlayers []*player.Entity, log *log.Entit
 	}
 
 	if (action == "steal" || action == "assassinate") && !player.Dead() && !challengeLoss && !targetedPlayer.Dead(){
-		if targetedPlayer.Strategy().GetBlock(log, table.PlayerNames(), table.PlayerCoins(), table.FaceupDecks(), player.Deck()) {
+		if targetedPlayer.Strategy().GetBlock(log, table.PlayerNames(), table.PlayerCoins(), table.FaceupDecks(), targetedPlayer.Deck()) {
 			log.CreateBlock()
 			var blockCardClaim int
 			if action == "steal" {
-				blockCardClaim = targetedPlayer.Strategy().GetStealBlockCardChoice(log, table.PlayerNames(), table.PlayerCoins(), table.FaceupDecks(), player.Deck())
+				blockCardClaim = targetedPlayer.Strategy().GetStealBlockCardChoice(log, table.PlayerNames(), table.PlayerCoins(), table.FaceupDecks(), targetedPlayer.Deck())
 				log.CreateBlockCardClaim(blockCardClaim)
+				if blockCardClaim != 2 && blockCardClaim != 5 {
+					disqualifyPlayer(targetedPlayer, table, log, "Block card '" + strconv.Itoa(blockCardClaim) + "' is not valid for blocking steal'")
+				}
 			}
 			if player.Strategy().GetChallenge(log, table.PlayerNames(), table.PlayerCoins(), table.FaceupDecks(), player.Deck()) {
 				var neededCardAction string
@@ -138,11 +152,24 @@ func DoTurn(player *player.Entity, otherPlayers []*player.Entity, log *log.Entit
 				} else if action == "assassinate" {
 					neededCardAction = "block"
 				}
-				challengeSuccess := !player.Deck().HasCardForAction(neededCardAction)
+				challengeSuccess := !targetedPlayer.Deck().HasCardForAction(neededCardAction)
 				losingPlayer := player
+				var cardValue int
 				if challengeSuccess {
 					losingPlayer = targetedPlayer
-					swapChallengedCard(player, table, log, 1)
+					switch neededCardAction {
+					case "tax":
+						cardValue = 1
+					case "steal":
+						cardValue = 2
+					case "assassinate":
+						cardValue = 3
+					case "block":
+						cardValue = 4
+					case "exchange":
+						cardValue = 5
+					}
+					swapChallengedCard(player, table, log, cardValue)
 				} else {
 					challengeLoss = true
 					losingPlayer = player
@@ -165,9 +192,9 @@ func DoTurn(player *player.Entity, otherPlayers []*player.Entity, log *log.Entit
 					challengeSuccess := !otherPlayers[i].Deck().HasCardForAction("tax")
 					losingPlayer := player
 					if challengeSuccess {
-						challengeLoss = true
 						losingPlayer = otherPlayers[i]
 					} else {
+						challengeLoss = true
 						losingPlayer = player
 						swapChallengedCard(otherPlayers[i], table, log, 1)
 					}
@@ -219,31 +246,31 @@ func DoTurn(player *player.Entity, otherPlayers []*player.Entity, log *log.Entit
 		c1 := table.CenterDeck().TakeCards(1)[0]
 		c2 := table.CenterDeck().TakeCards(1)[0]
 		player.Deck().GiveCards([]int{c1,c2})
-		r1,r2 := player.Strategy().GetExchangeKeepChoices(log, table.PlayerNames(), table.PlayerCoins(), table.FaceupDecks(), player.Deck())
+		r1,r2 := player.Strategy().GetExchangeReturnChoices(log, table.PlayerNames(), table.PlayerCoins(), table.FaceupDecks(), player.Deck())
 		if !player.Deck().HasCards(r1, r2) {
 			disqualifyPlayer(player, table, log, "Player gave invalid card returns " + strconv.Itoa(r1) + "," + strconv.Itoa(r2))
 			table.CenterDeck().GiveCards([]int{c1,c2})
 		} else {
 			fourCards := player.Deck().TakeCards(4)
 			returnToTableCards := make([]int, 2)
-			player.Deck().GiveCards([]int{r1,r2})
 			r1found := false
 			r2found := false
 			for i := range fourCards {
+				var index int
+				if returnToTableCards[1] == 0 {
+					index = 1
+				}
+				if returnToTableCards[0] == 0 {
+					index = 0
+				}
 				if fourCards[i] == r1 && !r1found{
 					r1found = true
+					returnToTableCards[index] = fourCards[i]
 				} else if fourCards[i] == r2 && !r2found{
 					r2found = true
-				} else {
-					var index int
-					if returnToTableCards[1] == 0 {
-						index = 1
-					}
-					if returnToTableCards[0] == 0 {
-						index = 0
-					}
 					returnToTableCards[index] = fourCards[i]
-					fmt.Printf("Returned %d\n", fourCards[i])
+				} else {
+					player.Deck().GiveCards([]int{fourCards[i]})
 				}
 			}
 			table.CenterDeck().GiveCards(returnToTableCards)
@@ -315,22 +342,26 @@ func revealCard(player *player.Entity, table *table.Entity, log *log.Entity) int
 	lossChoice := player.Strategy().GetLossChoice(log, table.PlayerNames(), table.PlayerCoins(), table.FaceupDecks(), player.Deck())
 	if lossChoice == 0 {
 		cardLoss = player.Deck().TakeBottomCard()
-	} else {
+	} else if lossChoice == 1 {
 		cardLoss = player.Deck().TakeTopCard()
+	} else {
+		disqualifyPlayer(player, table, log, "invalid logChoice value" + strconv.Itoa(lossChoice))
+		cardLoss = 0
 	}
-	losingPlayerFaceupDeck := table.FaceupDecks()[player.Name()]
-	if losingPlayerFaceupDeck[0] != 0 {
-		losingPlayerFaceupDeck[1] = losingPlayerFaceupDeck[0]
-	}
-	losingPlayerFaceupDeck[0] = cardLoss
-	if table.FaceupDecks()[player.Name()][0] != 0 && table.FaceupDecks()[player.Name()][1] != 0 {
-		player.Kill()
+	if !player.Dead() {
+		losingPlayerFaceupDeck := table.FaceupDecks()[player.Name()]
+		if losingPlayerFaceupDeck[0] != 0 {
+			losingPlayerFaceupDeck[1] = losingPlayerFaceupDeck[0]
+		}
+		losingPlayerFaceupDeck[0] = cardLoss
+		if table.FaceupDecks()[player.Name()][0] != 0 && table.FaceupDecks()[player.Name()][1] != 0 {
+			player.Kill()
+		}
 	}
 	return cardLoss
 }
 
 func swapChallengedCard(player *player.Entity, table *table.Entity, log *log.Entity, card int) {
-	fmt.Println("Swapping Card")
 	table.CenterDeck().GiveCards([]int{card})
 	table.CenterDeck().ShuffleCards()
 	player.Deck().ReplaceCardWithCard(card, table.CenterDeck().TakeCards(1)[0])
